create `entities` endpoint
